requirements.txt
typer~=0.7.0
setuptools~=60.2.0
PyYAML~=6.0
numpy~=1.24.2
scipy~=1.10.1
matplotlib~=3.7.1
colorama~=0.4.6
PyWavelets~=1.4.1